requirements.txt
Django==5.0
django-filter==23.5
django-jazzmin==2.6.0
djangorestframework==3.14.0
Pillow==10.1.0
python-dotenv==1.0.0
whitenoise==6.6.0